fix: inline markdown code rendering and split TextMessage
react-markdown v9+ no longer passes the `inline` prop, so every `<code>`
was falling through the `!inline` branch and rendering as a fenced
CodeBlock. Detect fenced blocks via the `language-*` className instead
and treat everything else as inline code.

Also split TextMessage into focused files: TextMessage.jsx stays as the
Markdown orchestrator while CopyButton, CodeBlock, and InlineCode move
into textMessage/.
