+++ DashAI/front/src/components/generative/textMessage/CopyButton.jsx
@@ -1,0 +1,31 @@
+import { useState } from "react";
+import { IconButton, Tooltip } from "@mui/material";
+import CheckIcon from "@mui/icons-material/Check";
+import ContentCopyIcon from "@mui/icons-material/ContentCopy";
+
+export function CopyButton({ text }) {
+  const [copied, setCopied] = useState(false);
+
+  const handleCopy = () => {
+    navigator.clipboard.writeText(text).then(() => {
+      setCopied(true);
+      setTimeout(() => setCopied(false), 2000);
+    });
+  };
+
+  return (
+    <Tooltip title={copied ? "Copied!" : "Copy"}>
+      <IconButton
+        size="small"
+        onClick={handleCopy}
+        sx={{ color: "grey.400", "&:hover": { color: "grey.100" } }}
+      >
+        {copied ? (
+          <CheckIcon fontSize="inherit" />
+        ) : (
+          <ContentCopyIcon fontSize="inherit" />
+        )}
+      </IconButton>
+    </Tooltip>
+  );
+}
